requirements.txt
absl-py==0.8.0
aiofiles==0.4.0
aiohttp==3.5.4
altgraph==0.16.1
appdirs==1.4.3
APScheduler==3.6.0
asgiref==3.2.3
askcli==1.0.4
asn1crypto==0.24.0
astor==0.8.0
astroid==2.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==19.1.0
Automat==0.7.0
awscli==1.16.283
backcall==0.2.0
beautifulsoup4==4.8.1
bleach==3.1.5
blis==0.4.1
boto3==1.9.146
botocore==1.13.19
bs4==0.0.1
bz2file==0.98
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.12.3
characteristic==14.3.0
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.14.0
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.0
distlib==0.3.0
dj-database-url==0.5.0
Django==2.1
django-allauth==0.41.0
django-cors-headers==3.2.1
django-countries==6.1.2
django-extensions==2.2.9
django-heroku==0.3.1
django-material==1.6.3
django-multiforloop==0.2.1
django-mysql==3.5.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
dopamine-rl==2.0.5
ebaysdk==2.2.0
en-core-web-md==2.2.0
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
fbmessenger==6.0.0
filelock==3.0.12
firebase-admin==4.2.0
Flask==1.1.1
Flask-Cors==3.0.8
future==0.17.1
gast==0.2.2
gcloud==0.18.3
gevent==1.4.0
gin-config==0.2.1
google-api-core==1.16.0
google-api-python-client==1.8.1
google-auth==1.14.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-speech==1.3.2
google-cloud-storage==1.26.0
google-cloud-texttospeech==1.0.1
google-pasta==0.1.7
google-resumable-media==0.5.0
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.23.0
gunicorn==19.7.1
gym==0.14.0
h11==0.8.1
h2==3.1.1
h5py==2.10.0
hpack==3.0.0
httpcore==0.3.0
httplib2==0.13.1
httptools==0.0.13
humanfriendly==4.18
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.0
importlib-resources==1.4.0
imutils==0.5.3
incremental==17.5.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.10.1
jmespath==0.9.4
jsonpickle==1.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
jws==0.1.3
kafka-python==1.4.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
kiwisolver==1.1.0
klein==19.6.0
lazy-object-proxy==1.3.1
lxml==4.4.1
macholib==1.11
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mattermostwrapper==2.1
mccabe==0.6.1
mesh-tensorflow==0.0.5
mistune==0.8.4
mod-wsgi==4.7.1
MouseInfo==0.1.3
mpmath==1.1.0
msgpack==1.0.0
multidict==4.5.2
murmurhash==1.0.2
mws==0.8.6
mysql-connector-python==8.0.6
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.3
nltk==3.4.5
notebook==6.0.3
numpy==1.18.4
oauth2client==4.1.2
oauthlib==3.1.0
opencv-python==4.1.1.26
openpyxl==3.0.3
opt-einsum==3.1.0
packaging==19.0
pandas==1.0.3
pandocfilters==1.4.2
parsel==1.5.2
parso==0.7.0
pathlib==1.0.1
pathlib2==2.3.5
pathtools==0.1.2
pdfkit==0.6.1
pefile==2019.4.18
pi==0.1.2
pickleshare==0.7.5
pika==1.0.1
Pillow==7.1.2
pipenv==2018.11.26
plac==0.9.6
preshed==3.0.2
prometheus-client==0.8.0
promise==2.2.1
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.11.2
psycopg2==2.7.1
psycopg2-binary==2.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycparser==2.19
pycryptodome==3.9.7
PyDispatcher==2.0.5
pydot==1.4.1
pydub==0.23.1
PyGetWindow==0.0.8
pyglet==1.3.2
Pygments==2.6.1
PyHamcrest==1.9.0
PyInstaller==3.6
PyJWT==1.7.1
pykwalify==1.7.0
pylint==2.2.2
pymongo==3.8.0
PyMsgBox==1.0.8
pynput==1.6.8
pyodbc==4.0.28
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyperclip==1.8.0
pypiwin32==223
pypng==0.0.20
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyqt5-tools==5.15.0.1.7
pyreadline==2.1
Pyrebase4==4.3.0
PyRect==0.1.4
pyrsistent==0.15.4
PyScreeze==0.1.26
PySocks==1.7.1
pytesseract==0.3.0
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-dotenv==0.10.3
python-engineio==3.9.3
python-jwt==2.0.1
python-socketio==4.3.1
python-telegram-bot==11.1.0
python-walmart==0.1.7
python3-openid==3.1.0
PyTweening==1.0.3
pytz==2020.1
pywin32==225
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
qtstyles==0.0.2
questionary==1.1.1
queuelib==1.5.0
rasa==1.4.0
rasa-nlu==0.13.2
rasa-sdk==1.4.0
redis==3.3.5
requests==2.23.0
requests-async==0.5.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.3.2
rocketchat-API==0.6.31
rsa==3.4.2
ruamel.yaml==0.15.94
s3transfer==0.2.1
sanic==19.9.0
Sanic-Cors==0.9.9.post1
sanic-jwt==1.3.1
Sanic-Plugins-Framework==0.8.2
schedule==0.6.0
scikit-learn==0.20.2
scipy==1.2.1
Scrapy==1.8.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==1.3.1
soupsieve==1.9.5
spacy==2.2.1
SQLAlchemy==1.3.8
sqlparse==0.3.0
srsly==0.1.0
sympy==1.4
tabulate==0.8.3
tensor2tensor==1.14.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-datasets==1.2.0
tensorflow-docs==0.0.0
tensorflow-estimator==2.0.1
tensorflow-gan==1.0.0.dev0
tensorflow-metadata==0.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
thinc==7.1.1
tornado==6.0.4
tqdm==4.31.0
traitlets==4.3.3
Tubes==0.2.0
twilio==6.26.3
Twisted==19.7.0
typed-ast==1.1.0
typing==3.7.4.1
typing-extensions==3.7.4
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.9
virtualenv==20.0.18
virtualenv-clone==0.5.3
w3lib==1.21.0
wasabi==0.2.2
watchdog==0.10.2
wcwidth==0.1.7
webencodings==0.5.1
webexteamssdk==1.1.1
websocket-client==0.54.0
websockets==8.0.2
Werkzeug==0.16.0
wfastcgi==3.0.0
whitenoise==3.3.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
yarl==1.3.0
zipp==3.1.0
zope.interface==4.6.0
